backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.27.0
boto3==1.34.0
redis==5.0.0
celery==5.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.0
alembic==1.13.0
runpod==1.6.0
websockets==12.0
aiofiles==23.2.0
Pillow==10.2.0
moviepy==1.0.3